Split main tests up into main, delta, bracket, int_norm, and ratios
This allows for better overview of all the tests, the main test file
started to become rather large.
